+++ Kernel/Net/RTL8139NetworkAdapter.h
@@ -41,7 +41,7 @@ public:
     RTL8139NetworkAdapter(PCI::Address, u8 irq);
     virtual ~RTL8139NetworkAdapter() override;
 
-    virtual void send_raw(const u8*, int) override;
+    virtual void send_raw(const u8*, size_t) override;
     virtual bool link_up() override { return m_link_up; }
 
 private:
